+++ shared/src/schemas/api/ExerciseEntries.api.zod.ts
@@ -69,7 +69,8 @@ export const exerciseEntrySetResponseSchema = z
     set_type: z.string().nullable(),
     reps: z.number().nullable(),
     weight: z.number().nullable(),
-    duration: z.number().nullable(),
+    // Per-set duration is integer SECONDS (issue #1903).
+    duration: z.number().int().nullable(),
     rest_time: z.number().nullable(),
     notes: z.string().nullable(),
     rpe: z.number().nullable(),
@@ -97,7 +98,8 @@ export const exerciseEntrySetRequestSchema = z
     set_type: z.string().nullable().optional(),
     reps: z.number().nullable().optional(),
     weight: z.number().nullable().optional(),
-    duration: z.number().nullable().optional(),
+    // Per-set duration is integer SECONDS (issue #1903).
+    duration: z.number().int().nullable().optional(),
     rest_time: z.number().nullable().optional(),
     notes: z.string().nullable().optional(),
     rpe: z.number().nullable().optional(),
